Fix USB CDC ECM FIFO Interrupt storm
Found that program gets stuck in PCD_WriteEmptyTxFifo when handling interrupts. Xfer_count was zero while the Xfer_len was greater than zero, this scenario was not being handled by the PCD_WriteEmptyTxFifo and therefore not clearing the fifo empty interrupt. Since xfer_count seems to be related to multi packet transfers it could be that support for multi packet transfers is not being enabled in the code which results in this issue however directions as to what configuration to change to enable multi packet transfers was not found in documentation.
CI: Richard Keelan